+++ selfdrive/car/honda/radar_interface.py
@@ -24,7 +24,7 @@ class RadarInterface(RadarInterfaceBase):
     self.radar_off_can = CP.radarOffCan
     self.radar_ts = CP.radarTimeStep
 
-    self.delay = round(0.1 / CP.radarTimeStep)   # 0.1s delay of radar
+    self.delay = int(round(0.1 / CP.radarTimeStep))   # 0.1s delay of radar
 
     # Nidec
     if self.radar_off_can:
