+++ selfdrive/car/subaru/carcontroller.py
@@ -22,11 +22,11 @@ class CarController():
     # *** steering ***
     if (frame % self.p.STEER_STEP) == 0:
 
-      apply_steer = round(actuators.steer * self.p.STEER_MAX)
+      apply_steer = int(round(actuators.steer * self.p.STEER_MAX))
 
       # limits due to driver torque
 
-      new_steer = round(apply_steer)
+      new_steer = int(round(apply_steer))
       apply_steer = apply_std_steer_torque_limits(new_steer, self.apply_steer_last, CS.out.steeringTorque, self.p)
       self.steer_rate_limited = new_steer != apply_steer
