+++ selfdrive/car/volkswagen/carcontroller.py
@@ -40,7 +40,7 @@ class CarController():
       # when exceeding ~1/3 the 360 second timer.
 
       if c.active and CS.out.vEgo > CS.CP.minSteerSpeed and not (CS.out.standstill or CS.out.steerError or CS.out.steerWarning):
-        new_steer = round(actuators.steer * P.STEER_MAX)
+        new_steer = int(round(actuators.steer * P.STEER_MAX))
         apply_steer = apply_std_steer_torque_limits(new_steer, self.apply_steer_last, CS.out.steeringTorque, P)
         self.steer_rate_limited = new_steer != apply_steer
         if apply_steer == 0:
