+++ selfdrive/thermald/thermald.py
@@ -280,9 +280,9 @@ def thermald_thread(end_event, hw_queue):
       pass
 
     msg.deviceState.freeSpacePercent = get_available_percent(default=100.0)
-    msg.deviceState.memoryUsagePercent = round(psutil.virtual_memory().percent)
-    msg.deviceState.cpuUsagePercent = [round(n) for n in psutil.cpu_percent(percpu=True)]
-    msg.deviceState.gpuUsagePercent = round(HARDWARE.get_gpu_usage_percent())
+    msg.deviceState.memoryUsagePercent = int(round(psutil.virtual_memory().percent))
+    msg.deviceState.cpuUsagePercent = [int(round(n)) for n in psutil.cpu_percent(percpu=True)]
+    msg.deviceState.gpuUsagePercent = int(round(HARDWARE.get_gpu_usage_percent()))
 
     msg.deviceState.networkType = last_hw_state.network_type
     msg.deviceState.networkStrength = last_hw_state.network_strength
